refactor : CommonImages 컴포넌트 추가된 것으로 내용 수정
