+++ src/components/App.js
@@ -1,8 +1,13 @@
 import React from 'react';
 import Landing from './landing/Landing';
+import Home from './home/Home';
+
+// function App() {
+//   return <Landing />;
+// }
 
 function App() {
-  return <Landing />;
+  return true ? <Home /> : <Landing />;
 }
 
 export default App;
