+++ src/components/CommonImages.js
@@ -1,4 +1,19 @@
 import LogoImg from '../assets/logo.svg';
 import TitleImg from '../assets/img_title.png';
+import SnsIconF from '../assets/snsIcon_f.svg';
+import SnsIconT from '../assets/snsIcon_t.svg';
+import SnsIconY from '../assets/snsIcon_y.svg';
+import SnsIconI from '../assets/snsIcon_i.svg';
+import snsIconG from '../assets/snsIcon_g.svg';
+import snsIconK from '../assets/snsIcon_k.svg';
 
-export { LogoImg, TitleImg };
+export {
+  LogoImg,
+  TitleImg,
+  SnsIconF,
+  SnsIconT,
+  SnsIconY,
+  SnsIconI,
+  snsIconG,
+  snsIconK,
+};
