--- src/components/Login/Images.js
@@ -1,6 +1,0 @@
-import eyeOffImg from '../../assets/eye-off.svg';
-import eyeOnImg from '../../assets/eye-on.svg';
-import snsIconG from '../../assets/snsIcon_g.svg';
-import snsIconK from '../../assets/snsIcon_k.svg';
-
-export { eyeOffImg, eyeOnImg, snsIconG, snsIconK };
