+++ src/components/home/style.css
@@ -1,0 +1,196 @@
+html {
+    font-family: Pretendard, sans-serif;
+    word-break: keep-all;
+    min-height: 100%;
+}
+
+body {
+    margin: 0;
+    background-image:
+            linear-gradient(180deg, #f0f6ff 20%, #111322 20%);
+    display:flex;
+    flex-direction: column;
+    height: 100%;
+    min-height: 100%;
+}
+
+/* header */
+header{
+    background-color: #F0F6FF;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    width:100%;
+}
+
+header nav{
+    padding: 1.25rem 12.5rem 2rem;
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    position: fixed;
+    top:0;
+    z-index: 1;
+    background-image: linear-gradient(
+            180deg,
+            rgba(240, 246, 255, 1) 70%,
+            rgba(240, 246, 255, 0)
+    );
+    width: 100%;
+    max-width: 192rem;
+}
+
+a{
+    display: flex;
+    align-items: center;
+    text-decoration: none;
+}
+
+/* hero-header*/
+.hero-header{
+    background-color: #f0f6ff;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    text-align: center;
+    /*margin-bottom: 7.5rem;*/
+    padding: 4.38rem  0 0;
+    margin-top: 7rem;
+    width: 100%;
+}
+
+.hero-header h1{
+    font-size: 4rem;
+    font-style: normal;
+    font-weight: 700;
+    line-height: 5rem;
+    margin: 0;
+
+}
+
+.hero-header h1 span{
+    background: linear-gradient(91deg, #6D6AFE 17.28%, #FF9F9F 74.98%);
+}
+
+.btn-add-link{
+    width: 21.875rem;
+    margin: 2.5rem;
+}
+
+.hero-header img{
+    width: 69.875rem;
+    height: 41.2rem;
+}
+
+
+/* section */
+section {
+    display: flex;
+    padding: 3.12rem 10rem;
+    gap: 9.81rem;
+    background-color: #fff;
+    justify-content: center;
+    align-items: center;
+}
+
+section.save{
+    padding-top: 7.5rem;
+}
+
+section:nth-child(even){
+    flex-direction: row-reverse;
+}
+
+.section-text{
+    display: flex;
+    flex-direction: column;
+}
+
+.section-text h2{
+    font-size: 3rem;
+    font-style: normal;
+    font-weight: 700;
+    line-height: normal;
+    letter-spacing: -0.01875rem;
+    margin: 0;
+}
+
+.section-text p{
+    color: #6B6B6B;
+    font-size: 1rem;
+    font-style: normal;
+    font-weight: 500;
+    margin: 0;
+}
+
+section img{
+    width: 34.375rem;
+}
+
+/* section span 그라데이션 배경색 */
+.save span{
+    background: linear-gradient(96deg, #FE8A8A 1.72%, #A4CEFF 74.97%);
+}
+
+.manage span{
+    background: linear-gradient(277deg, #6FBAFF 59.22%, #FFD88B 93.66%);
+}
+
+.share span{
+    background: linear-gradient(99deg, #6D7CCD 19.76%, rgba(82, 136, 133, 0.22) 52.69%);
+}
+
+.search span{
+    background: linear-gradient(271deg, #FE578F -9.84%, #68E8F9 107.18%);
+}
+
+.section-text span,
+.hero-header h1 span{
+    background-clip: text;
+    -webkit-background-clip: text;
+    -webkit-text-fill-color: transparent;
+}
+
+
+/* footer */
+footer{
+    background-color: #111322;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+}
+
+.footer-container{
+    display: flex;
+    padding: 2rem 6.5rem 4rem;
+    justify-content: space-between;
+    height: 10rem;
+    width: 100%;
+    max-width: 192rem;
+}
+
+footer span{
+    color: #676767;
+    text-align: center;
+    font-size: 1.6rem;
+    font-weight: 400;
+    display: flex;
+    align-items: center;
+}
+
+.footer-menu{
+    display: flex;
+    gap: 1.88rem;
+    font-size: 1.6rem;
+}
+
+.footer-menu a{
+    color: #CFCFCF;
+}
+
+.snsIcons{
+    display: flex;
+    gap: 0.75rem;
+    color: #fff;
+}
