+++ src/components/login/Images.js
@@ -1,0 +1,4 @@
+import eyeOffImg from '../../assets/eye-off.svg';
+import eyeOnImg from '../../assets/eye-on.svg';
+
+export { eyeOffImg, eyeOnImg };
