+++ src/components/login/SignUp.js
@@ -1,11 +1,8 @@
 import React from 'react';
-import '../../reset.css';
-import '../../common_style.css';
-import './style.css';
-import { LogoImg } from '../CommonImages';
-import { eyeOffImg, snsIconG, snsIconK } from './Images';
+import { LogoImg, snsIconG, snsIconK } from '../CommonImages';
+import { eyeOffImg } from './Images';
 
-function SignIn() {
+function SignUp() {
   return (
     <>
       <div className="container">
@@ -16,9 +13,9 @@ function SignIn() {
             </a>
           </div>
           <nav>
-            <span>회원이 아니신가요?</span>
-            <a className="nav-link" href="/pages/login/signup.html">
-              회원 가입하기
+            <span>이미 회원이신가요?</span>
+            <a className="nav-link" href="/pages/login/signin.html">
+              로그인 하기
             </a>
           </nav>
         </header>
@@ -38,16 +35,25 @@ function SignIn() {
                 alt="비밀번호 보이게하기(현재 보이지 않음)"
               />
             </div>
+            <div className="form-item">
+              <label htmlFor="password-check">비밀번호 확인</label>
+              <input id="password-check" type="password" name="password" />
+              <img
+                className="eye"
+                src={eyeOffImg}
+                alt="비밀번호 보이게하기(현재 보이지 않음)"
+              />
+            </div>
           </form>
-          <div className="btn">로그인</div>
+          <div className="btn">회원가입</div>
           <div className="sns">
-            <span>소셜 로그인</span>
+            <span>다른 방식으로 가입하기</span>
             <div className="sns-icons">
               <a href="https://www.google.com/">
-                <img src={snsIconG} alt="구글 로그인 바로가기" />
+                <img src={snsIconG} alt="구글 회원가입 바로가기" />
               </a>
               <a href="https://www.kakaocorp.com/page/">
-                <img src={snsIconK} alt="카카오 로그인 바로가기" />
+                <img src={snsIconK} alt="카카오 회원가입 바로가기" />
               </a>
             </div>
           </div>
@@ -57,4 +63,4 @@ function SignIn() {
   );
 }
 
-export default SignIn;
+export default SignUp;
